+++ ak-py/src/agentkernel/api/__init__.py
@@ -11,5 +11,5 @@ try:
 except importlib.metadata.PackageNotFoundError:
     __version__ = "0.1.0"
 
-from .rest_request_handler import RESTRequestHandler
-from .restapi import RESTAPI
+from .handler import RESTRequestHandler
+from .http import RESTAPI
