--- ak-py/src/agentkernel/api/rest_request_handler.py
@@ -1,26 +1,0 @@
-from abc import ABC, abstractmethod
-
-from fastapi import APIRouter
-
-
-class RESTRequestHandler(ABC):
-    @abstractmethod
-    def get_router(self) -> APIRouter:
-        """
-        Returns the APIRouter instance which has configured routes
-        E.g.:
-        - GET /health: Health check
-        - GET /agents: List available agents
-
-        router = APIRouter()
-
-        @router.get("/health")
-        def health():
-            return {"status": "ok"}
-
-        @router.get("/agents")
-        def list_agents():
-            return {"agents": list(GlobalRuntime.instance().agents().keys())}
-
-        """
-        pass
