+++ ak-py/src/agentkernel/deployment/aws/aklambda.py
@@ -4,7 +4,7 @@ import logging
 import traceback
 from typing import Any, Dict
 
-from ..core import AgentService
+from ...core import AgentService
 
 logging.basicConfig(
     level=logging.DEBUG,
