+++ ak-py/src/agentkernel/framework/openai/openai.py
