+++ ak-py/src/agentkernel/mcp.py
@@ -1,0 +1,8 @@
+import importlib.metadata
+
+try:
+    __version__ = importlib.metadata.version("agentkernel")
+except importlib.metadata.PackageNotFoundError:
+    __version__ = "0.1.0"
+
+from .api.mcp import *
